% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_genetics.R
\name{align_pheno_geno}
\alias{align_pheno_geno}
\title{Align Phenotype and Genotype Data}
\usage{
align_pheno_geno(pheno, geno, id_col = "Genotype")
}
\arguments{
\item{pheno}{Dataframe containing phenotypic data.}

\item{geno}{GRM matrix.}

\item{id_col}{Character string. Name of the genotype column in pheno.}
}
\value{
A list containing aligned pheno and geno.
}
\description{
Ensures intersection of genotypes between phenotypic data and GRM.
}
