% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsd.R
\name{amsd.multinomial}
\alias{amsd.multinomial}
\title{Absolute Mean Standardized Differences (multinomial)}
\usage{
amsd.multinomial(DT, v, t, v_xy = NULL, na.rm = TRUE, calc_var = FALSE)
}
\arguments{
\item{DT}{a class \code{'data.table'} object}

\item{v}{character scalar with the name of the variable to be calculated}

\item{t}{character scalar with the name of the treatment variable}

\item{v_xy}{Expected \code{list(var_x, var_y)}, for pre-computed variances (ie. scaling factors)
to be used in AMSD calculation. The scaling factors should be the same pre and post matching
for comparability.}

\item{calc_var}{Logical. Do you wish to calculate pooled variance?}
}
\value{
A numeric scalar
}
\description{
calculate absolute mean standardized differences between vectors X and Y.
with \code{calc_var= TRUE} can be used to calcualte variances vs return AMSD. Useful for
pre-calculating the scaling factor \code{v_xy}
}
