% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amsd.R
\name{amsd.numeric}
\alias{amsd.numeric}
\title{Absolute Mean Standardized Differences (numeric)}
\usage{
amsd.numeric(x, y, v_xy = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{\code{numeric} vector, assumed to be "treated" units}

\item{y}{\code{numeric} vector, assumed to be "control" units}

\item{v_xy}{Expected \code{list(var_x, var_y)}, for pre-computed variances (ie. scaling factors)
to be used in AMSD calculation. The scaling factors should be the same pre and post matching
for comparability.}
}
\value{
A numeric scalar
}
\description{
calculate absolute mean standardized differences between vectors X and Y
}
