% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_model.r
\name{audit_model_terms}
\alias{audit_model_terms}
\title{Audit ASReml Model for Design Validation}
\usage{
audit_model_terms(model)
}
\arguments{
\item{model}{A fitted \code{asreml} object.}
}
\value{
Prints diagnostic messages. Returns TRUE if pass, FALSE if warnings found.
}
\description{
Performs a "Design Tableau" audit on a fitted ASReml model.
It checks if the model structure respects the physical trial design (Row/Column).
Failing to account for spatial variation in field trials is a common source of error.
}
