% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avPlots.invis.R
\name{avPlots.invis}
\alias{avPlots.invis}
\title{A function that extracts the information from the avPlots function}
\usage{
avPlots.invis(MODEL, ...)
}
\arguments{
\item{model}{This is the model you want to check linearity for}
}
\value{
'ggAVplots()' return av plots for the variables in the model
}
\description{
A function that extracts the information from the avPlots function
}
