% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_cross_variance_cpp}
\alias{calc_cross_variance_cpp}
\title{Calculate Gametic Variance for a Cross (Biparental)}
\usage{
calc_cross_variance_cpp(p1, p2, effects, map, max_dist = 3)
}
\arguments{
\item{p1}{NumericVector. Genotype of Parent 1 (0 or 2).}

\item{p2}{NumericVector. Genotype of Parent 2 (0 or 2).}

\item{effects}{NumericVector. Marker effects (a).}

\item{map}{NumericVector. Genetic map positions in Morgans.}

\item{max_dist}{double. Maximum distance (in Morgans) to calculate
covariance. Markers further apart are assumed unlinked. Default 3.0.}
}
\value{
double. The gametic variance.
}
\description{
Computes the expected genetic variance in the F2 population derived from
crossing two inbred parents, accounting for linkage disequilibrium.
}
