% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envirotyping.R
\name{calculate_days_since_start}
\alias{calculate_days_since_start}
\title{Calculate Days Since Start}
\usage{
calculate_days_since_start(doy_vector, planting_date_string)
}
\arguments{
\item{doy_vector}{Numeric vector of DOY values.}

\item{planting_date_string}{Character string format "YYYY-Month-DD" (e.g., "2023-October-04").}
}
\value{
Numeric vector of days elapsed since planting.
}
\description{
Converts a vector of Day of Year (DOY) values into "Days Since Start" based on a planting date string.
Handles year transitions (e.g., planting in Oct, sampling in Jan).
}
