% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fa.R
\name{calculate_fast_indices}
\alias{calculate_fast_indices}
\title{Calculate FAST Indices (OP and RMSD)}
\usage{
calculate_fast_indices(fa_object, k = NULL)
}
\arguments{
\item{fa_object}{An object of class `fa_model` or a list with `loadings` and `scores`.}

\item{k}{Number of factors.}
}
\value{
A dataframe with Genotype, OP, and RMSD.
}
\description{
Calculates "Factor Analytic Selection Tools" indices from a FA Model.
}
