% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_h2_from_list}
\alias{calculate_h2_from_list}
\title{Calculate Heritability from List}
\usage{
calculate_h2_from_list(model_list, id = "Genotype")
}
\arguments{
\item{model_list}{List of ASReml objects.}

\item{id}{Character. The genetic term name (default "Genotype").}
}
\value{
A named vector of H2 values.
}
\description{
Robustly calculates heritability for a list of ASReml models.
}
