% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fa.R
\name{calculate_i_classes}
\alias{calculate_i_classes}
\title{Determine Interaction Classes (iClasses)}
\usage{
calculate_i_classes(fa_object, alpha = 0.05)
}
\arguments{
\item{fa_object}{An FA model object.}

\item{alpha}{Significance level for specific interactions (approx).}
}
\value{
Dataframe with iClass columns.
}
\description{
Classifies genotypes based on their interaction scores (Significant vs NS).
}
