% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fa.R
\name{calculate_op}
\alias{calculate_op}
\title{Calculate Overall Performance (OP) and RMSD}
\usage{
calculate_op(model)
}
\arguments{
\item{model}{An object of class \code{fa_model}.}
}
\value{
A \code{fast_selection} object.
}
\description{
Calculates OP (Factor 1 performance) and RMSD (deviation/instability) and returns a selection object.
}
