% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_spatial.R
\name{calculate_reliability}
\alias{calculate_reliability}
\title{Calculate Reliability from Model}
\usage{
calculate_reliability(model, genotype)
}
\description{
Calculate Reliability from Model
}
\keyword{internal}
