% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fa.R
\name{calculate_rmsd}
\alias{calculate_rmsd}
\title{Calculate RMSD (Wrapper)}
\usage{
calculate_rmsd(model)
}
\arguments{
\item{model}{An object of class \code{fa_model}.}
}
\value{
A \code{fast_selection} object (same as calculate_op).
}
\description{
Calculate RMSD (Wrapper)
}
