% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate model under/over fit}
\usage{
calibrate(DT, var1, var2, y, yhat = "hat_y")
}
\arguments{
\item{DT}{A \code{data.table} object with your data}

\item{var1}{A character scalar for the name of \eqn{X_1} in \code{DT} to condition on.}

\item{var2}{A character scalar for the name of \eqn{X_2} in \code{DT}  to condition on.}

\item{y}{A character scalar for the name of \eqn{Y} in \code{DT}}

\item{yhat}{A character scalar for the name of \eqn{\hat Y} in \code{DT}}
}
\value{
A class \code{data.table} object
}
\description{
Calibrate model predictions conditional on a two-dimensional grid of predictors.
Specifically, check if the model is over (or under) predicting vs true rate.
\eqn{Cal_{j,k} = frac{\sum_{i=1}^N \hat Y_i}{\sum_{i=1}^N Y_i} | X_j = x_j, X_k = x_k}
}
