% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asreml.R
\name{check_asreml_availability}
\alias{check_asreml_availability}
\title{Check ASReml Availability}
\usage{
check_asreml_availability(action = "stop")
}
\arguments{
\item{action}{Character. "stop" (default) or "warning" or "logical".}
}
\value{
Logical TRUE if available, or error/warning if not.
}
\description{
Checks if the `asreml` package is installed and licensed.
Stops execution with a clear message if unavailable.
}
