% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_spatial.R
\name{check_design_terms}
\alias{check_design_terms}
\title{Check Experimental Design Factors by Study}
\usage{
check_design_terms(
  data,
  study_col = "studyName",
  row_col = "rowNumber",
  col_col = "colNumber",
  block_col = "blockNumber",
  rep_col = "replicate"
)
}
\arguments{
\item{data}{The dataframe containing the MET data.}

\item{study_col}{String. Name of the column defining the Study/Experiment (default "studyName").}

\item{row_col}{String. Name of the Row column (default "rowNumber").}

\item{col_col}{String. Name of the Column column (default "colNumber").}

\item{block_col}{String. Name of the Block column (default "blockNumber").}

\item{rep_col}{String. Name of the Replicate column (default "replicate").}
}
\value{
A dataframe summarizing which terms are available for each study.
}
\description{
Scans a MET dataset to determine which design terms (Row, Column, Block, Replicate)
are valid for inclusion in a model for each specific study.
}
