% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_connectivity.R
\name{check_met_connectivity}
\alias{check_met_connectivity}
\title{Check MET Connectivity (Island Detection)}
\usage{
check_met_connectivity(
  data,
  trial_col = "Trial",
  gen_col = "Genotype",
  stop_if_disconnected = FALSE
)
}
\arguments{
\item{data}{Dataframe containing the MET data.}

\item{trial_col}{Character. Column name for Trial/Environment.}

\item{gen_col}{Character. Column name for Genotype.}

\item{stop_if_disconnected}{Logical. If TRUE, stops execution if islands are found.}
}
\value{
A list containing:
  \item{n_clusters}{Integer. Number of disconnected components.}
  \item{membership}{Dataframe mapping Trial to Cluster.}
  \item{islands}{Character vector of Trial names that are disconnected from the main group.}
}
\description{
Scans a Multi-Environment Trial (MET) dataset to identify disconnected clusters
of environments. Disconnected trials ("islands") will result in singular
factor analytic models.
}
