% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_spatial.R
\name{compute_weights_from_model}
\alias{compute_weights_from_model}
\title{Compute Stage 2 Weights form Model}
\usage{
compute_weights_from_model(model, genotype, threshold = 1e-08)
}
\description{
Compute Stage 2 Weights form Model
}
\keyword{internal}
