% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHeader.R
\name{CreateHeader}
\alias{CreateHeader}
\title{A function that takes the first row of a dataframe and makes it the header}
\usage{
CreateHeader(dataframe)
}
\arguments{
\item{dataframe}{This is a dataframe}
}
\value{
'CreateHeader()' returns a dataframe where the previous first row has now become the header.
}
\description{
A function that takes the first row of a dataframe and makes it the header
}
\examples{
df <- data.frame(x = c("Height",180), y = c("Weight", 160))
df
CreateHeader(df)
}
