% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{DEFAULT_CHECKS}
\alias{DEFAULT_CHECKS}
\title{Default Checks}
\format{
An object of class \code{character} of length 6.
}
\usage{
DEFAULT_CHECKS
}
\description{
Standard check genotypes for plots.
}
\keyword{datasets}
