% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_design.R
\name{diagnose_design}
\alias{diagnose_design}
\title{Diagnose Trial Design Structure}
\usage{
diagnose_design(data, genotype, trial, rep = NULL)
}
\arguments{
\item{data}{Dataframe containing the trial data.}

\item{genotype}{Character. Column name for genotype.}

\item{trial}{Character. Column name for trial.}

\item{rep}{Character. Optional column name for rep.}
}
\description{
Check using `diagnose_design(data, "Geno", "env")`.
}
