% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fa.R
\name{extract_fa_blups}
\alias{extract_fa_blups}
\title{Extract Factor Analytic BLUPs and Scores}
\usage{
extract_fa_blups(fa_object)
}
\arguments{
\item{fa_object}{An FA model object.}
}
\value{
A list containing the raw/rotated scores.
}
\description{
Extract Factor Analytic BLUPs and Scores
}
