% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fa.R
\name{extract_fa_loadings}
\alias{extract_fa_loadings}
\title{Extract Factor Analytic Loadings and VAF}
\usage{
extract_fa_loadings(fa_object)
}
\arguments{
\item{fa_object}{An FA model object.}
}
\value{
A list containing the rotated loadings matrix and the Variance Accounted For (VAF) summary.
}
\description{
Extract Factor Analytic Loadings and VAF
}
