% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_tree.R
\name{fac_tree}
\alias{fac_tree}
\title{A function that takes number and returns its base factors}
\usage{
fac_tree(x)
}
\arguments{
\item{number}{any number greater than 0}
}
\value{
'fac_tree' returns a vector containing all the base facotrs of the input
}
\description{
A function that takes number and returns its base factors
}
\examples{
fac_tree(12)
}
