% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fa.R, R/fa_selection.R
\name{fast_tools}
\alias{fast_tools}
\title{Factor Analytic Selection Tools (FAST)}
\value{
A \code{fast_selection} object or dataframes of indices.

A \code{fast_selection} object or dataframes of indices.
}
\description{
Functions to calculate Smith & Cullis (2018) selection indices from a Factor Analytic model.

Functions to calculate Smith & Cullis (2018) selection indices from a Factor Analytic model.
}
