% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fa.R
\name{fit_fa_model}
\alias{fit_fa_model}
\title{Extract and Rotate Factor Analytic Model Parameters}
\usage{
fit_fa_model(model, classify, psi_term = NULL, rotate = TRUE)
}
\arguments{
\item{model}{A fitted object of class \code{asreml}.}

\item{classify}{Character. The term to extract. E.g., \code{"fa(Site, 2):Genotype"}.}

\item{psi_term}{Character (Optional). For RR models, the term containing specific variances.
E.g., \code{"diag(Site):Genotype"}. If NULL for RR, attempts automatic detection.}

\item{rotate}{Logical. Perform SVD rotation to PC solution? Default TRUE.}
}
\value{
An object of class \code{fa_model}.
}
\description{
The primary engine of the toolkit. This function parses a fitted ASReml-R model object to
extract loadings, scores (BLUPs), and specific variances from a Factor Analytic (FA)
or Reduced Rank (RR) model term.
}
\details{
It performs a Singular Value Decomposition (SVD) rotation to align the solution
with the Principal Component axis. This rotation is required to validly calculate
the **Factor Analytic Selection Tools (FAST)** indices (Overall Performance and Stability)
as proposed by Smith & Cullis (2018).
}
