% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_spatial.R
\name{fit_site_model}
\alias{fit_site_model}
\title{Fit Single Site Model (Internal Helper)}
\usage{
fit_site_model(
  sub_data,
  trait,
  genotype,
  block,
  row,
  col,
  resids_list,
  compute_weights = FALSE,
  weight_threshold = 0.01,
  workspace = "1gb"
)
}
\description{
Fits a spatial model for a single environment. Used by \code{analyze_single_trial}.
}
\keyword{internal}
