% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FA_asreml_toolkit.r
\name{generate_grm}
\alias{generate_grm}
\title{Generate Synthetic Genomic Relationship Matrix (GRM)}
\usage{
generate_grm(n_gen = 200, n_markers = 1000, seed = 123)
}
\description{
Generate Synthetic Genomic Relationship Matrix (GRM)
}
