% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicFlow-package.R
\docType{package}
\name{GenomicFlow-package}
\alias{GenomicFlow}
\alias{GenomicFlow-package}
\title{GenomicFlow: Genomic Selection and Factor Analytic Tools for Plant Breeding}
\description{
A comprehensive suite for genomic selection and trial analysis. Implements the "Design Tableau" for trial validation, wrapping ASReml-R for Factor Analytic models, and providing Factor Analytic Selection Tools (FAST) as per Smith & Cullis (2018). It also includes population development tools utilizing gametic variance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zaja10/GenomicFlow}
  \item Report bugs at \url{https://github.com/zaja10/GenomicFlow/issues}
}

}
\author{
\strong{Maintainer}: Zachary Aldiss \email{zaldiss@illinois.edu}

}
\keyword{internal}
