% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.r
\name{get_i_classes}
\alias{get_i_classes}
\title{Calculate Interaction Classes (iClasses)}
\usage{
get_i_classes(object, factor = 2, threshold = 0.1)
}
\arguments{
\item{object}{An object of class \code{fa_asreml}.}

\item{factor}{Integer. Which factor defines the interaction? (Default 2).}

\item{threshold}{Numeric. The loading magnitude required to assign a site to a class
(e.g., 0.1). Sites between -0.1 and 0.1 are considered "Neutral".}
}
\value{
A list containing:
\describe{
  \item{site_classes}{Dataframe assigning each site to Class_Pos, Class_Neg, or Neutral.}
  \item{gen_effects}{Dataframe with predicted genetic values for each genotype in
  the Positive vs Negative environments, and the Differential (Crossover).}
}
}
\description{
Implements the methodology of Smith et al. (2021) to detect Specific Adaptation.
It partitions environments into "Positive" and "Negative" classes based on the
sign and magnitude of their loadings for a specific Factor (usually Factor 2).
}
\details{
Predictions are calculated as:
\deqn{Pred_{Class} = (Score_{F1} \times \bar{\lambda}_{F1}) + (Score_{Fk} \times \bar{\lambda}_{Class})}
This combines the variety's general performance (F1) with its specific response to the
interaction driver (Fk).
}
