% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAVplots.R
\name{ggAVplots}
\alias{ggAVplots}
\title{A function that takes a linear model and returns av plots that are ggplot objects}
\usage{
ggAVplots(MODEL, color = "red")
}
\arguments{
\item{color}{This is the color of the dotted line}

\item{model}{This is the model you want to check linearity for}
}
\value{
'ggAVplots()' return av plots for the variables in the model
}
\description{
A function that takes a linear model and returns av plots that are ggplot objects
}
