% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_design.R
\name{makeNm}
\alias{makeNm}
\title{Convert Factor to Numeric safely}
\usage{
makeNm(df)
}
\arguments{
\item{df}{Vector or column to convert.}
}
\value{
Numeric vector.
}
\description{
Convert Factor to Numeric safely
}
