% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_spatial.R
\name{mat2sparse}
\alias{mat2sparse}
\title{Convert Symmetric Matrix to Sparse Three-Column Format}
\usage{
mat2sparse(mat)
}
\arguments{
\item{mat}{A symmetric numeric matrix.}
}
\value{
A dataframe with columns "Row", "Col", "Value".
}
\description{
Converts a symmetric weight matrix (Inverse PEV) into the "Row, Col, Value" format
required by ASReml's \code{vm()} function.
}
\keyword{internal}
