% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_small_factor_levels.R
\name{merge_small_factor_levels}
\alias{merge_small_factor_levels}
\title{Merge small factor levels.}
\usage{
merge_small_factor_levels(x, min.perc = 0.01, case_weights = NULL)
}
\arguments{
\item{x}{Either character of factor vector}
}
\value{
Factor vector, with small levels merged
}
\description{
Given an input factor/character vector,
  merges together all vector levels that account for
  less than \code{min.perc} of the total number of
  elements in the vector
}
