% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asreml.R
\name{mkConv}
\alias{mkConv}
\title{Force Convergence by Variance Change}
\usage{
mkConv(mod)
}
\arguments{
\item{mod}{asreml object.}
}
\value{
Converged asreml object.
}
\description{
Iteratively updates the model until the percentage change in variance components
is below a threshold (default 1%).
}
