% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_stat.R
\name{mode_stat}
\alias{mode_stat}
\title{Calculate a vector mode value}
\usage{
mode_stat(x)
}
\arguments{
\item{x}{An atomic vector.}
}
\value{
The mode value of x.
}
\description{
Calculate a vector mode value.
}
