% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_t_test.R
\name{pairwise_t_test}
\alias{pairwise_t_test}
\title{Pairwise T Test}
\usage{
pairwise_t_test(
  y_bar,
  sigma,
  n,
  p.adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none"),
  welch = FALSE
)
}
\arguments{
\item{y_bar}{A \code{numeric} vector of sample sizes}

\item{sigma}{A \code{numeric} vector of associated sample standard deviations}

\item{p.adjust.method}{A \code{character} string specifying the correcton method. See
\code{\link[stats]{p.adjust}}.}

\item{welch}{Logical. Do you want to calculate a Welch T-test with Welch-Satterthwaite
degrees of freedom?}
}
\value{
an object of class 'pairwise.htest'
}
\description{
Pairwise T-test for independent two-sample tests  with assumed approximately
equal variances. Also allows Welch-Satterthwaite DF. Can also be used for single T-tests
vs multiple pairwise comparisons
}
\examples{
{
    ## single t-test
    pairwise_t_test(y_bar= rnorm(2), sigma= rnorm(2, mean=1)
                    , n= sample.int(100, size= 2), p.adjust.method= 'none')
    ## pairwise tests
    Y <- rnorm(10); sig <- rnorm(10, mean= 1); n <- sample.int(100, size= 10)
    pairwise_t_test(y_bar= Y, sigma= sig, n= n, p.adjust.method= 'holm')
}
}
