% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design_tableau.R
\name{plot.design_tableau}
\alias{plot.design_tableau}
\title{Plot Design Tableau}
\usage{
\method{plot}{design_tableau}(x, ...)
}
\arguments{
\item{x}{A design_tableau object.}

\item{...}{Additional arguments.}
}
\description{
Visualizes the randomization design using the first two variables in the plot formula
(e.g., Site and Rep) to show coverage/balance.
}
