% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_biometrics.R
\name{plot.fa_asreml}
\alias{plot.fa_asreml}
\title{Master Visualization Suite for GenomicFlow}
\usage{
\method{plot}{fa_asreml}(x, type = "fast", factor = NULL, n_label = 5, highlight = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{fa_asreml}.}

\item{type}{A character string specifying the visualization module.}

\item{factor}{Integer or Vector. Controls which factors are visualized.}

\item{n_label}{Integer. Number of top genotypes to label.}

\item{highlight}{Character vector. Names of specific genotypes to highlight.}

\item{...}{Additional graphical parameters.}
}
\description{
The central plotting interface for the toolkit. Dispatches to specific visualization engines.
}
