% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_selection.R
\name{plot.fast_selection}
\alias{plot.fast_selection}
\title{Plot FAST Selection}
\usage{
\method{plot}{fast_selection}(x, type = "op_rmsd", ...)
}
\arguments{
\item{x}{A \code{fast_selection} object.}

\item{type}{Character. "op_rmsd" (default).}

\item{...}{Additional arguments.}
}
\description{
Plot FAST Selection
}
