% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_gs.R
\name{plot.gs_predictions}
\alias{plot.gs_predictions}
\title{Plot Method for Genomic Selection Predictions}
\usage{
\method{plot}{gs_predictions}(x, ...)
}
\arguments{
\item{x}{An object of class `gs_predictions`.}

\item{...}{Additional arguments (ignored).}
}
\value{
Returns the `patchwork` combined figure.
}
\description{
Plot Method for Genomic Selection Predictions
}
