% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.h2_comparison}
\alias{plot.h2_comparison}
\title{Plot Heritability/Reliability Comparison}
\usage{
\method{plot}{h2_comparison}(x, ...)
}
\arguments{
\item{x}{An object of class \code{h2_comparison}.}

\item{...}{Additional arguments.}
}
\description{
Plot Heritability/Reliability Comparison
}
