% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.mt_selection_results}
\alias{plot.mt_selection_results}
\title{Plot Selection Pipeline Results}
\usage{
\method{plot}{mt_selection_results}(x, type = "profile", ...)
}
\arguments{
\item{x}{An object of class \code{mt_selection_results}.}

\item{type}{Character. "profile" for the Selection Profile (Parallel Coordinates),
"gain" for the Response to Selection bar chart. Default is "profile".}

\item{...}{Additional arguments.}
}
\description{
Generates visualizations for the Multi-Trait Selection Pipeline.
}
