% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_trial_layout.R
\name{plot.padded_trial}
\alias{plot.padded_trial}
\title{Plot Trial Layout}
\usage{
\method{plot}{padded_trial}(x, fill_col = NULL, ...)
}
\arguments{
\item{x}{Object of class `padded_trial`.}

\item{fill_col}{Character. Column to map to fill color.}

\item{...}{Unused.}
}
\description{
Usage: plot(padded_object, fill_col = "Yield")
}
