% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbarts_pdp.R
\name{plot.pdp}
\alias{plot.pdp}
\title{Partial Dependency Plot}
\usage{
\method{plot}{pdp}(pdp, xind, plquants = c(0.05, 0.95), rng_y = c(0.7, 1), factor = FALSE)
}
\arguments{
\item{pdp}{A class \code{dbarts::pdbart} object. The result of \code{\link[dbarts]{pdbart}}.}

\item{xind}{Integer scalar indicating which variable to be plotted.}

\item{plquants}{Numeric vector of length two. In pdp plots, beliefs about \eqn{f(x)} are
indicated by plotting the posterior median, lower, and upper quantiles. \code{plquants} indicates
the upper and lower quantiles to be displayed.}

\item{rng_y}{Numeric vector of length two. Used to limit the range of the Y variable in the plot.}
}
\value{
A ggplot plot object
}
\description{
Create a partial dependence plot using \code{library(ggplot2)} instead of base R's
graphics.
}
\seealso{
\code{\link[dbarts]{pdbart}}.
}
