% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_diagnostics.r
\name{plot_connectivity}
\alias{plot_connectivity}
\title{Visualise MET Connectivity}
\usage{
plot_connectivity(
  data,
  x = "Year",
  y = NULL,
  trace = "Genotype",
  method = "count",
  ...
)
}
\arguments{
\item{data}{A data.frame.}

\item{x}{String. Factor for X-axis (e.g. "Year" or "Trial").}

\item{y}{String. Factor for Y-axis (Default `NULL` uses x for symmetric).}

\item{trace}{String. Variable establishing connection (e.g. "Genotype").}

\item{method}{String. "count", "jaccard", or "prop_min".}

\item{...}{Additional arguments passed to `image()`.}
}
\description{
Visualizes the connectivity structure of a MET dataset using a heatmap.
Identify if trials or years are connected by common germplasm.
}
