% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_gain.R
\name{plot_genetic_trend}
\alias{plot_genetic_trend}
\title{Plot Genetic Trend (Era Plot)}
\usage{
plot_genetic_trend(
  data,
  year_col = "Year",
  value_col = "OP",
  check_list = NULL,
  ...
)
}
\arguments{
\item{data}{A dataframe containing genotype performance.}

\item{year_col}{String. Column for Year of Release.}

\item{value_col}{String. Column for Breeding Value / OP.}

\item{check_list}{Character vector. Names of checks to highlight.}

\item{...}{Additional arguments to plot.}
}
\description{
Visualizes the genetic trend over time, highlighting checks and the regression line.
}
