% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_response}
\alias{plot_response}
\title{Plot Predicted Response to Selection}
\usage{
plot_response(resp_df)
}
\arguments{
\item{resp_df}{Dataframe returned by \code{predict_response}.}
}
\description{
Visualizes the percentage change in traits if the top selection is advanced.
}
