% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_spatial}
\alias{plot_spatial}
\title{Plot Spatial Map}
\usage{
plot_spatial(input, row = "Row", col = "Column", fill = "Yield")
}
\arguments{
\item{input}{Either a dataframe or `asreml` object.}

\item{row}{Colname for Row.}

\item{col}{Colname for Column.}

\item{fill}{Colname for value to plot. If input is model, "Residuals" is used.}
}
\description{
Visualizes a field trial as a heatmap.
Wraps ASReml residuals or raw data.
}
