% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_pre_analysis.R, R/plotting.R
\name{plot_trial_map}
\alias{plot_trial_map}
\title{Flexible Trial Map}
\usage{
plot_trial_map(
  data,
  trial_val = NULL,
  trial_col = "Trial",
  row = "Row",
  col = "Column",
  val = "Yield"
)

plot_trial_map(
  data,
  trial_val = NULL,
  trial_col = "Trial",
  row = "Row",
  col = "Column",
  val = "Yield"
)
}
\arguments{
\item{data}{A data.frame.}

\item{trial_val}{Value. The specific trial ID to filter.}

\item{trial_col}{String. Column name for Trial ID.}

\item{row}{String. Column name for Row.}

\item{col}{String. Column name for Column.}

\item{val}{String. Variable to fill heatmap (e.g. "Yield").}
}
\description{
Plots the spatial layout of a specific trial. Adapts to missing spatial coordinates
by falling back to a 1D visualization if necessary.
}
