% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{predict_gain}
\alias{predict_gain}
\title{Predict Genetic Gain}
\usage{
predict_gain(object, selection_percent = 5)
}
\arguments{
\item{object}{An object of class \code{fa_model}.}

\item{selection_percent}{Numeric. Top percentage to select (e.g., 5 for 5\%).}
}
\value{
A dataframe of predicted gains per environment/trait.
}
\description{
Estimates expected genetic gain (R) using the Breeder's Equation: \eqn{R = i r \sigma_g}.
Useful for evaluating the ROI of the breeding cycle.
}
