% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictgls.R
\name{predictgls}
\alias{predictgls}
\title{A function that returns the standardized residuals form a gls object}
\usage{
predictgls(glsobj, newdframe = NULL, level = 0.95)
}
\arguments{
\item{newdframe}{A dataframe of new values you want to predict on}

\item{level}{The confidence level you want around your predictions}

\item{model}{gls object}
}
\value{
'predictgls()' returns predicted values using a gls object
}
\description{
A function that returns the standardized residuals form a gls object
}
