% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_genetics.R
\name{prepare_asreml_grm}
\alias{prepare_asreml_grm}
\title{Prepare GRM for ASReml (Inverse & Sparse)}
\usage{
prepare_asreml_grm(G, blend = 0.02)
}
\arguments{
\item{G}{A symmetric numeric matrix (GRM).}

\item{blend}{Numeric. Amount of identity matrix to blend (0-1) to ensure invertibility.
Default 0.02 (2% bending).}
}
\value{
A matrix with attribute 'INVERSE' = TRUE.
}
\description{
Checks a Genomic Relationship Matrix (GRM) for singularity, bends it if necessary,
inverts it, and converts it to the sparse 'giv' format required by ASReml.
}
