% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_genetics.R
\name{prepare_trial_grm}
\alias{prepare_trial_grm}
\title{Prepare Trial GRM}
\usage{
prepare_trial_grm(master_grm, trial_data, genotype_col = "Genotype")
}
\arguments{
\item{master_grm}{A square numeric matrix representing the master GRM.}

\item{trial_data}{A dataframe containing the trial data.}

\item{genotype_col}{Character. Column name in `trial_data` containing the genotypes.}
}
\value{
A square matrix representing the trial-specific GRM, scaled and padded.
}
\description{
Subsets a master Genomic Relationship Matrix (GRM) to the lines present in a trial.
Automatically pads the matrix with a dummy identity matrix for ungenotyped lines
to prevent ASReml singularity/inverse errors.
}
