% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_number.R
\name{prime}
\alias{prime}
\title{A function that takes number and returns if its prime or not}
\usage{
prime(n)
}
\arguments{
\item{number}{any number greater than 0}
}
\value{
'prime()' returns true or false depending on if the input was prime or not.
}
\description{
A function that takes number and returns if its prime or not
}
\examples{
prime(12)
}
