% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design_tableau.R
\name{print.design_tableau}
\alias{print.design_tableau}
\title{Print Design Tableau}
\usage{
\method{print}{design_tableau}(x, ...)
}
\arguments{
\item{x}{A design_tableau object.}

\item{...}{Additional arguments.}
}
\description{
Print Design Tableau
}
