% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fa.R
\name{print.fa_model}
\alias{print.fa_model}
\title{Print Method for Factor Analytic Models}
\usage{
\method{print}{fa_model}(x, ...)
}
\arguments{
\item{x}{An object of class `fa_model`.}

\item{...}{Additional arguments (ignored).}
}
\value{
Prints a concise summary and returns the object invisibly.
}
\description{
Print Method for Factor Analytic Models
}
