% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_gs.R
\name{print.gs_predictions}
\alias{print.gs_predictions}
\title{Print Method for Genomic Selection Predictions}
\usage{
\method{print}{gs_predictions}(x, ...)
}
\arguments{
\item{x}{An object of class `gs_predictions`.}

\item{...}{Additional arguments (ignored).}
}
\value{
Prints a summary and returns the object invisibly.
}
\description{
Print Method for Genomic Selection Predictions
}
