% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote_gen.R
\docType{data}
\name{quotes}
\alias{quotes}
\title{A function that returns a quote}
\format{
An object of class \code{character} of length 12.
}
\usage{
quotes
}
\value{
'quote_gen()' returns a quote
}
\description{
A function that returns a quote
}
\examples{
quote_gen()
}
\keyword{datasets}
