% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fa.R
\name{rank_genotypes}
\alias{rank_genotypes}
\title{Rank Genotypes using Selection Index}
\usage{
rank_genotypes(fast_df, weight_stability = 1)
}
\arguments{
\item{fast_df}{Dataframe containing OP and RMSD columns.}

\item{weight_stability}{Numeric. penalty 'b'.}
}
\value{
The dataframe sorted by Index.
}
\description{
Ranks genotypes based on a Smith-Cullis index: I = OP - b * RMSD.
}
