% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parallel.R
\name{run_parallel}
\alias{run_parallel}
\title{Run Function in Parallel (OS Agnostic)}
\usage{
run_parallel(X, FUN, cores = NULL, ...)
}
\arguments{
\item{X}{Vector or List to iterate over.}

\item{FUN}{Function to apply.}

\item{cores}{Number of cores.}

\item{...}{Args for FUN.}
}
\description{
Run Function in Parallel (OS Agnostic)
}
