% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_spatial.R
\name{select_spatial_model}
\alias{select_spatial_model}
\title{Select Best Spatial Model}
\usage{
select_spatial_model(
  sub_data,
  trait,
  genotype,
  block,
  resids_list,
  workspace = "1gb"
)
}
\description{
Select Best Spatial Model
}
\keyword{internal}
