% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdres.gls.R
\name{stdres.gls}
\alias{stdres.gls}
\title{A function that returns the standardized residuals form a gls object}
\usage{
stdres.gls(glsobj)
}
\arguments{
\item{model}{This is the model you want standardized residuals for}
}
\value{
'stdres.gls()' returns the decorrelated, adjusted, standardized residuals for a gls model
}
\description{
A function that returns the standardized residuals form a gls object
}
