% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions-package.R
\docType{package}
\name{UtilityFunctions-package}
\alias{UtilityFunctions}
\alias{UtilityFunctions-package}
\title{UtilityFunctions: Factor Analytic and Design Tools}
\description{
A suite of tools for processing and visualizing Factor Analytic models
in plant breeding (ASReml-R).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zaja10/UtilityFunctions}
  \item Report bugs at \url{https://github.com/zaja10/UtilityFunctions/issues}
}

}
\author{
\strong{Maintainer}: Zachary Aldiss \email{zaldiss@illinois.edu}

}
\keyword{internal}
