% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_estimators.R
\name{var_est}
\alias{var_est}
\title{Unequal Probability Sampling. Estimate variance}
\usage{
var_est(
  y,
  pi,
  mu,
  na.rm = TRUE,
  trim = TRUE,
  trim_max = 100,
  method = c("Horvitz-Thomson", "Generalized-unequal-prob", "Hansen-Hurwitz",
    "weighted-mean")
)
}
\arguments{
\item{y}{A numeric vector of outcomes for all units}

\item{pi}{A numeric vector  of inclusion probabilities for each unit.}

\item{mu}{A numeric scalar for the mean estimate. See \code{\link{mean_est}}.}

\item{trim}{Logical. Should extreme weights be trimmed?}

\item{trim_max}{Numeric scalar for the max relative weight to trim. Defaults to \code{100}}

\item{method}{One of \code{c('Horvitz-Thomson', 'Generalized-unequal-prob', 'Hansen-Hurwitz', 'weighted-mean')}.
See Details.}
}
\description{
Mean estimators for unequal probability samples. See details.
}
\section{Details}{

Multiple methods are implemented. Hansen-Hurwitz is the only provided sampling with replacement
estimator. Weighted mean is also provided but produces a biased estimator. In general,
'Hansen-Hurwitz' and 'Horvitz-Thompson' are preferred if \code{y} is approximately proportional
to \code{pi}; 'Generalized-unequal-prob' is preferred if \code{y} and \code{pi} are not well
related. No estimator will be uniformly better than any other

Note: We assume independence for all units \eqn{i, j, \in \{1, \ldots, N \}} if \eqn{i \neq j}.
}

