% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_cross_variance_cpp}
\alias{calc_cross_variance_cpp}
\title{Calculate Gametic Variance for a Cross (Biparental)}
\usage{
calc_cross_variance_cpp(p1, p2, effects, map)
}
\arguments{
\item{p1}{NumericVector. Genotype of Parent 1 (0 or 2).}

\item{p2}{NumericVector. Genotype of Parent 2 (0 or 2).}

\item{effects}{NumericVector. Marker effects (a).}

\item{map}{NumericVector. Genetic map positions in Morgans.}
}
\value{
double. The gametic variance.
}
\description{
Computes the expected genetic variance in the F2 population derived from crossing
two inbred parents, accounting for linkage disequilibrium.
}
\details{
Variance = Sum(Var_i) + 2 * Sum(Cov_ij)
Var_i = a_i^2 / 2  if segregating (P1 != P2)
Cov_ij = (1 - 2r_ij) * (a_i * a_j) / 2
where r_ij is recombination frequency calculated from Haldane map function:
r_ij = 0.5 * (1 - exp(-2 * |d_i - d_j|))

Assumes parents are fully homozygous inbred lines.
}
