% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fa.R
\name{calculate_d_optimality}
\alias{calculate_d_optimality}
\title{Calculate D-Optimality (Network Efficiency)}
\usage{
calculate_d_optimality(object)
}
\arguments{
\item{object}{An object of class \code{fa_model}.}
}
\value{
A list containing total_d and site_impact.
}
\description{
Quantifies the information content of the trial network.
}
