% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fa.R, R/fa_selection.R
\name{calculate_i_classes}
\alias{calculate_i_classes}
\title{Calculate Interaction Classes (iClasses)}
\usage{
calculate_i_classes(object, factor = 2, threshold = 0.1)

calculate_i_classes(object, factor = 2, threshold = 0.1)
}
\arguments{
\item{object}{An object of class \code{fa_asreml}.}

\item{factor}{Integer. Which factor defines the interaction? (Default 2).}

\item{threshold}{Numeric. Loading magnitude for class assignment.}
}
\value{
A list with site classes and genetic effects.

A list with site classes and genetic effects.
}
\description{
Implements Smith et al. (2021) to detect Specific Adaptation.

Implements Smith et al. (2021) to detect Specific Adaptation.
}
