% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_i_classes}
\alias{calculate_i_classes}
\title{Calculate Interaction Classes (Generalized)}
\usage{
calculate_i_classes(object)
}
\arguments{
\item{object}{An object of class \code{fa_model}.}
}
\value{
A list with site classes and genotype classes.
}
\description{
Classifies environments/genotypes based on the sign of their Factor Loadings/Scores.
Supports any number of factors (k).
}
