% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_fast_selection.R
\name{fast_tools}
\alias{fast_tools}
\alias{calculate_op}
\alias{calculate_rmsd}
\title{Factor Analytic Selection Tools (FAST)}
\usage{
calculate_op(model)

calculate_rmsd(model)
}
\arguments{
\item{model}{An object of class \code{fa_asreml}.}
}
\value{
A \code{fast_selection} object containing the selection table and metadata.
}
\description{
Functions to calculate Smith & Cullis (2018) selection indices from a Factor Analytic model.
}
\section{Functions}{
\itemize{
\item \code{calculate_op()}: Calculates OP and RMSD and returns a selection object.

\item \code{calculate_rmsd()}: Calculates RMSD specifically (currently integrated in calculate_op).

}}
