% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_design.R
\name{find_missing_plots}
\alias{find_missing_plots}
\title{Find Missing Plots}
\usage{
find_missing_plots(
  data,
  experiment = "Experiment",
  row = "Row",
  col = "Column"
)
}
\description{
Identifies gaps in rectangular grids (Row x Col).
}
