% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asreml.R
\name{force_convergence}
\alias{force_convergence}
\title{Force Convergence}
\usage{
force_convergence(model, max_iter = 10)
}
\arguments{
\item{model}{asreml object.}

\item{max_iter}{Numeric limit.}
}
\description{
Repeatedly updates an ASReml model until convergence or max iterations.
}
