% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{genetic_metrics}
\alias{genetic_metrics}
\title{Genetic Metrics Suite}
\description{
A unified collection of functions for calculating heritability, reliabilities,
and genetic gain from ASReml models and phenotypic data.
}
