% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FA_asreml_toolkit.r
\name{get_d_optimality}
\alias{get_d_optimality}
\title{Calculate D-Optimality (Network Efficiency)}
\usage{
get_d_optimality(object)
}
\arguments{
\item{object}{An object of class \code{fa_asreml}.}
}
\value{
A list containing:
\describe{
  \item{total_d}{The total determinant of the information matrix.}
  \item{site_impact}{A dataframe ranking sites by their \% contribution to network information.}
}
}
\description{
Quantifies the information content of the trial network based on the Factor Analytic
loadings ($\Lambda$). Implements the D-Optimality criterion to identify redundant sites.
}
