% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FA_asreml_toolkit.r
\name{get_i_classes}
\alias{get_i_classes}
\title{Calculate Interaction Classes (iClasses)}
\usage{
get_i_classes(object, factor = 2, threshold = 0.1)
}
\arguments{
\item{object}{An object of class \code{fa_asreml}.}

\item{factor}{Integer. Which factor defines the interaction? (Default 2).}

\item{threshold}{Numeric. The loading threshold to assign a site to a class
(e.g., loadings > 0.1 are Positive Class).}
}
\value{
A list containing site classifications and predicted genetic values for
        each genotype in the Positive vs Negative classes.
}
\description{
Implements the methodology of Smith et al. (2021) to partition environments into
biological classes based on their response to specific GxE drivers (Factors).
}
