% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.fa_model}
\alias{plot.fa_model}
\title{Master Visualization Suite for Factor Analytic Models}
\usage{
\method{plot}{fa_model}(x, type = "fast", factor = NULL, n_label = 5, highlight = NULL, ...)
}
\description{
Master Visualization Suite for Factor Analytic Models
}
