% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_h2.r
\name{plot.h2_comparison}
\alias{plot.h2_comparison}
\title{Plot Heritability/Reliability Comparison}
\usage{
\method{plot}{h2_comparison}(x, ...)
}
\arguments{
\item{x}{An object of class \code{h2_comparison} produced by \code{compare_h2()}.}

\item{...}{Additional arguments passed to \code{barplot}.}
}
\description{
Visualizes the output of \code{compare_h2} using a grouped bar chart.
Compares different metrics (Cullis, Oakey, Standard) across sites.
}
