% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_genomics.R
\name{predict_cross_utility}
\alias{predict_cross_utility}
\title{Predict Cross Utility (Genetic Variance)}
\usage{
predict_cross_utility(parents, effects, map, max_dist = 3)
}
\arguments{
\item{parents}{Matrix (2 x M) of parental genotypes (scored 0, 1, 2).}

\item{effects}{Vector (M) of marker effects.}

\item{map}{Vector (M) of genetic map positions (in Morgans).}

\item{max_dist}{Numeric. Maximum distance (Morgans) to consider linkage.}
}
\value{
A numeric value representing the expected variance.
}
\description{
Calculates the expected genetic variance of a biparental cross given
parental genotypes and marker effects.
}
