% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{predict_gain}
\alias{predict_gain}
\title{Predict Genetic Gain (Breeder's Equation)}
\usage{
predict_gain(selection_intensity = 1.755, reliability, genetic_variance)
}
\arguments{
\item{selection_intensity}{Numeric. Standardized selection intensity ($i$). Default 1.755 (10%% selected).}

\item{reliability}{Numeric. Reliability/Accuracy of the selection index ($r^2$ or $h^2$). If accuracy ($r$) is desired, use sqrt(rel).}

\item{genetic_variance}{Numeric. Additive genetic variance ($\sigma_g^2$).}
}
\value{
Numeric. Expected gain in units of the trait.
}
\description{
Estimates expected genetic gain ($R$) using the Breeder's Equation:
\eqn{R = i \times r \times \sigma_g}
}
