% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design_tableau.R, R/utils_design.R
\name{print.design_diagnosis}
\alias{print.design_diagnosis}
\title{Print Design Diagnosis}
\usage{
\method{print}{design_diagnosis}(x, ...)

\method{print}{design_diagnosis}(x, ...)
}
\arguments{
\item{x}{A design_diagnosis object.}

\item{...}{Additional arguments.}
}
\description{
Print Design Diagnosis

Print Design Diagnosis
}
