% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{structure_df}
\alias{structure_df}
\title{data.frame structure in data.frame format}
\usage{
structure_df(data, max_string_length = 60)
}
\arguments{
\item{data}{A data.frame for which the structure is to be given}

\item{max_string_length}{maximum string length in the sample_values column}
}
\value{
A data.frame with the following columns:
  1. variable: vairable name
  2. class: Variable class
  3. sample_values: Sample of the values that variable can take
    for factor variables and main stasitics (mean, median etc) for
    a numeric variable
  4. missing: Percent missing values
}
\description{
This function produces output very similar to
  the one obtained by the \code{str} function, only it returns
  a data.frame object which can be used for example in
  rmarkdown reports
}
