% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions-package.R
\docType{package}
\name{UtilityFunctions-package}
\alias{UtilityFunctions}
\alias{UtilityFunctions-package}
\title{UtilityFunctions: Factor Analytic Model Extraction and Selection Tools}
\description{
A specialized suite for processing factor analytic models. Implements "FAST" selection tools, design diagnostics, and cross prediction utilities involving gametic variance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zaja10/UtilityFunctions}
  \item Report bugs at \url{https://github.com/zaja10/UtilityFunctions/issues}
}

}
\author{
\strong{Maintainer}: Zachary Aldiss \email{zaldiss@illinois.edu}

}
\keyword{internal}
