% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{validate_met_data}
\alias{validate_met_data}
\title{Validate MET Data Structure}
\usage{
validate_met_data(data, trait, genotype, env, check_numeric = TRUE)
}
\description{
Performs robust quality checks on a Multi-Environment Trial dataset.
}
