% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wizard_Scorecard.R
\name{wizard}
\alias{wizard}
\title{A function that walks you through keeping score for a wizard game and returns the scorecard.}
\usage{
wizard(save = FALSE)
}
\arguments{
\item{save}{Is a logical that indicates whether you want to save the scorecard as a csv}
}
\value{
'wizard()' walks you through keeping score for a wizard game and returns the scorecard.
}
\description{
A function that walks you through keeping score for a wizard game and returns the scorecard.
}
\examples{
df <- wizard()
}
